+++ ChangeLog.rst
@@ -1,5 +1,5 @@
-HEAD
-====
+0.9 - 2023-10-02
+================
  - Bugfix: :class:`modelcif.SoftwareGroup` now allows for parameters to
    be associated with each piece of software in the group, rather than
    with the group as a whole (#33).